% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathToClip.R
\name{pathToClip}
\alias{pathToClip}
\title{Path To Clipboard}
\usage{
pathToClip()
}
\description{
RStudio Addin -- Copy Active Document's Path to Clipboard
}
\details{
On *nix systems, the clipr package is required and relies on
xclip or xsel. See for instance: https://bit.ly/2QEfrOB
}
