% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathToClip.R
\name{pathToClip_back}
\alias{pathToClip_back}
\title{Path To Clipboard (\\)}
\usage{
pathToClip_back()
}
\description{
RStudio Addin -- Copy Active Document's Path to Clipboard
}
\details{
Uses backslash as path separator.
}
