% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathToClip.R
\name{pathToClip_dbl_back}
\alias{pathToClip_dbl_back}
\title{Path To Clipboard (\\\\)}
\usage{
pathToClip_dbl_back()
}
\description{
RStudio Addin -- Copy Active Document's Path to Clipboard
}
\details{
Uses double backslash as path separator.
}
