% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathToClip.R
\name{pathToClip_dbs}
\alias{pathToClip_dbs}
\title{Path To Clipboard (\\\\)}
\usage{
pathToClip_dbs()
}
\description{
RStudio Addin -- Copy Active Document's Path to Clipboard
}
\details{
Produces double-backslash, Windows-style paths.
}
