% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathToClip.R
\name{pathToClip_fwd}
\alias{pathToClip_fwd}
\title{Path To Clipboard (/)}
\usage{
pathToClip_fwd()
}
\description{
RStudio Addin -- Copy Active Document's Path to Clipboard
}
\details{
Uses forward slash as path separator.
}
